+++ src/utils.ts
@@ -100,7 +100,7 @@ import { NEON_DEFAULT_DATABASE_NAME } from './constants.js';
 
 /**
  * Returns the default database for a project branch
- * If a database name is provided, it returns that name
+ * If a database name is provided, it fetches and returns that database
  * Otherwise, it looks for a database named 'neondb' and returns that
  * If 'neondb' doesn't exist, it returns the first available database
  * Throws an error if no databases are found
@@ -113,9 +113,7 @@ export async function getDefaultDatabase({
   projectId: string;
   branchId: string;
   databaseName?: string;
-}): Promise<string> {
-  if (databaseName) return databaseName;
-
+}) {
   const { data } = await neonClient.listProjectBranchDatabases(
     projectId,
     branchId,
@@ -125,8 +123,15 @@ export async function getDefaultDatabase({
     throw new Error('No databases found in your project branch');
   }
 
+  if (databaseName) {
+    const requestedDatabase = databases.find((db) => db.name === databaseName);
+    if (requestedDatabase) {
+      return requestedDatabase;
+    }
+  }
+
   const defaultDatabase = databases.find(
     (db) => db.name === NEON_DEFAULT_DATABASE_NAME,
   );
-  return defaultDatabase ? defaultDatabase.name : databases[0].name;
+  return defaultDatabase || databases[0];
 }